Move StartingGridEntry into RaceData.
