--- replayenhancer/StartingGridEntry.py
@@ -1,8 +1,0 @@
-class StartingGridEntry:
-    """
-    Represents an entry on the starting grid.
-    """
-    def __init__(self, position, driver_index, driver_name):
-        self.position = position
-        self.driver_index = driver_index
-        self.driver_name = driver_name
